refactor(app): migrate multipart list and copy-part flows
